+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 
-import { KakaoScript } from '@/components/kakaoScript';
+import { KakaoScript } from '@/components/KakaoScript';
 
 import ReactQueryProviders from '../providers/ReactQueryProvider';
