+++ src/components/KakaoScript/index.ts
